+++ packages/client-ledger/RELEASES.md
@@ -37,3 +37,13 @@ Based on:
 - [typescript v1.0.0-alpha.3] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.3] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.3 - packages/client-ledger
+
+## 2024-07-11 00:16:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.331.0 (2.365.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.4] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.4] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.4 - packages/client-ledger